Moved WorkspaceCompaniesList and CompanyProfilesList out from TableList-directory.
